created calculator

+++ group4Colab/src/main/java/com/sparta/Subtraction.java
@@ -1,7 +1,7 @@
 package com.sparta;
 
 public class Subtraction {
-    public static int operate(int number1,int number2){
+    public static double operate(double number1,double number2){
         return number1-number2;
 
     }
